backend/requirements.txt

fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.4
pydantic-settings==2.2.1
python-multipart==0.0.9
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.5
respx==0.20.2
numpy==1.26.4
PyPDF2==3.0.1
pdfminer.six==20231228
pypdfium2==4.25.0
pytesseract==0.3.10
Pillow==10.2.0
